Fix usage of `std::uniform_int_distribution` with unsupported types.
Some compilers don't like unsigned or less than 32-bit types.

PiperOrigin-RevId: 865045487

+++ test/simd/s16-simd-scalar.cc
@@ -32,7 +32,7 @@ class S16SimdSCALARTest : public ::testing::Test {
   void SetUp() override {
     inputs_.resize(3 * xnn_simd_size_s16);
     output_.resize(xnn_simd_size_s16);
-    std::uniform_int_distribution<int16_t> s16(-100, 100);
+    std::uniform_int_distribution<int> s16(-100, 100);
     std::generate(inputs_.begin(), inputs_.end(),
                   [&]() { return s16(rng_); });
   }
